Merged PR 19884: Fix out of memory error when saving segments with large customer database
## What's being changed
This PR changes the way we flag contacts for reimport when they fall within a segment being created in a store with a large customer database.

## Why it's being changed
Saving a segment containing huge numbers of customers was causing a fatal error (out of memory).

## How to review / test this change
- In M2 EE, use performance fixtures to generate a large number of customers
- Create a customer segment and save it
- Confirm that no fatal error is reported, and that the associated contact rows in `email_contact` are flagged for reimport (`email_imported = 0`)

## Notes
Although this issue was reported when working with a customer database of 2000000 records, you do not need to generate that many with the performance fixtures. The issue arose from using a collection to get all of the customer IDs within the segment, as a variable, then passing that to the contact resource to update the `email_contact` table.

This change simply removes that functionality and replaces it with a subquery based on the segment table the customer IDs are written to.

Related work items: #109849