Merged PR 26392: Update Page Builder MFTF tests
## What's being changed

In this PR we have amended our MFTF tests for Page Builder.

## Why it's being changed

In 2.4.2 certain elements - including dotdigital forms - cannot be dropped directly onto the stage. Instead they have to be placed into a container element such as a row, column or tab.

We have reverted to using the `ResetDotmailerAccountData` action group to reset the account creds, in favour of the newer `ResetAccountDataCLI` one. I don't have really robust evidence to back this up, but from running the suite many times, it seemed that the CLI action group wasn't reliably ensuring that the account state was how it should be at the start of the subsequent test. This is important with the `dotdigitalCommerce` suite because some of the tests need an active account, and the others are specifically testing an inactive account.

## How to review / test this change

- run `vendor/bin/mftf run:group dotdigitalCommerce`

Related work items: #132806