Merged PR 26552: Add merge patch for content types
## What's being changed

We are overriding a step in the core action group `validateMenuSectionsAndContentTypes`.

## Why it's being changed

A core test fails because it expects a certain number of menu items under the 'Add content' submenu - and because of our extension, it sees an extra item.

## How to review / test this change
```
vendor/bin/mftf run:test PageBuilderFullScreenLeftPanelHeightAndFunctionalityTest -r
```

## Notes

If Selenium times out, comment out the `<before>` and `<after>` blocks in the `PageBuilderWithWYSIWYG` suite.

Related work items: #133332