Merged PR 27313: Add plugin to return entity ids for grouped parents
## What's being changed

We've added a plugin to fetch product entity ids for linked row ids, for grouped product children.

## Why it's being changed

In Magento Commerce, there is a `row_id` column on the `catalog_product_entity` table that is not present in community edition. This column supports the scheduling of content updates for products.

Whereas in community, `catalog_product_link` has a foreign key to the `entity_id` column of `catalog_product_entity`, in commerce the `row_id` is the linked column. This has caused unexpected issues for merchants getting invalid parent ids for their grouped product children (i.e. incorrect URLs and image paths).

This plugin takes the list of parent row ids, and swaps them with the matching entity ids.

## How to review / test this change

- Create a new grouped product, and give it an image
- Create a new simple product, visibility = Not Visible Individually, no image this time
- In catalog_product_entity you should see two new rows
- Now add the simple product to the group
- Observe the relation in catalog_product_link
- Now schedule an update for the group product, any update, some time in the future
- Observe the additional rows in catalog_product_entity and catalog_product_link
- Run catalog sync

Related work items: #136065