Merged PR 27172: Fetch reward points balance from magento_reward table
## What's being changed

We're changing where we obtain our number for a customer's reward points balance.

## Why it's being changed

The `magento_reward_history` table has a `points_balance` column, but apparently this does not truly reflect a customer's remaining points _after a reward has expired_.

If a customer earns a reward worth 100 points, but this reward expires, merchants would expect the balance to be 0. This PR ensures that we sync the expected value.

## How to review / test this change

- Turn on reward points in Stores > Configuration > Customer > Reward Points
- Add points for customer registration and newsletter sign up
- Set the rewards to expire after 1 day
- Create a new customer account and sign up to the newsletter
- Check your reward points in the Magento admin and run contact sync
- Now wait for the rewards to expire
- Re-sync your customer
- The synced REWARD_POINTS value should match the figure in `magento_reward`

Related work items: #135558