Merged PR 33619: Fix popover display on mobile
## What's being changed

We've updated the `data-mobile` property to output true/false instead of 1/0.

## Why it's being changed

1/0 worked when this script was introduced (although apparently it was always wrong) but upstream changes mean it doesn't any longer. This change means that Dotdigital form content types added in popover mode will load in mobile viewports.

Merchants who currently have 'Show on mobile devices' set to Yes will need to re-save any existing Page Builder components to update the display.

Merchants who currently have 'Show on mobile devices' set to No will not be affected.

Note that we still say display on mobile devices is not recommended, but that is to do with UX best practice, and not to do with modals actually working or not on smaller screens.

## How to review / test this change

- Start with a popover form set to Show on mobile devices > No
- Switch to this branch
- Clear cache
- Reload the CMS page
- Ensure Show on mobile devices is still set to No
- Set Show on mobile devices to Yes, save and exit
- On the front end, use Chrome developer tools to display a mobile viewport (iPhone 6/7/8 fx)
- Reload the page
- See the popover (note that if you've already dismissed this popover, you'll need to manually delete the cookie '<form-id>-Dismissed' before reloading

Related work items: #159621