Merged PR 38572: Rename MFTF tests
## What's being changed

Most of our MFTF tests in this module.

## Why it's being changed

The updated tests had filenames not ending in 'Test', which is the convention, and also the filenames did not match the test name. Some of the tests only worked for me locally when I renamed them - I'm not sure how they worked before.

## How to review / test this change

- See pipeline MFTF build complete (green)

Related work items: #180388