Merged PR 40099: Set email_catalog constraint to disabled
## What's being changed

We've added a disabled foreign key to the module.

## Why it's being changed

Our Email module adds a foreign key constraint to our `email_catalog` table that prevents adding product ids to it that don't already exist in `catalog_product_entity`.

In Adobe Commerce, our foreign key is replaced with a different constraint: `EMAIL_CATALOG_PRODUCT_ID_SEQUENCE_PRODUCT_SEQUENCE_VALUE`. Commerce does things a bit differently, and ties product entities back to a `sequence_product` table.

What this means is that, when running `bin/magento setup:db:status`, in Commerce, it always says we need to upgrade our db - because there are constraints in the aggregated db_schema instructions that are not strictly reflected in the actual database schema.

Per the suggestion, because this only happens in Commerce, we can add a db_schema.xml in the Enterprise module that sets the constraint to disabled. Once this is added the warning goes away.

Fixes https://github.com/dotmailer/dotmailer-magento2-extension/issues/589.

## How to review / test this change

- On develop branch, run `bin/magento setup:db:status` to see the warning we're trying to fix
- Switch to this branch
- Run `bin/magento setup:upgrade`
- Run `bin/magento setup:db:status`, don't see a warning
- Try to add a row to email_catalog with a product_id that doesn't exist, make sure you still get a foreign key violation

Related work items: #184854