Merged PR 40425: Update Page Builder key to fix multiple appearances
## What's being changed

The key for 'appearance' has been updated to 'appearances' and applied to the script url path.

## Why it's being changed

The form key did not match the form data declaration setting the value to `default as string literal` which was an invalid value as the form expected an int for appearances, this caused the form to not respect the max appearances value set by the merchant.

## How to review / test this change

1. Add a new popover form to a page ( if its an old form resave the form)
2. Set 'Stop Displaying' to 'After a number of appearances' and change form to a 'popover'
3. set value from 1 to 2
4. save page.
5. On page check your cookies for `{form-id}-Dismissed` if the value if equal or higher than the max appearances the form should not show.
6. clear cookies and dismiss the form till its `{form-id}-Dismissed` value stops the form from displaying.

Related work items: #188296