Merged PR 42087: Add specific selector for dotdigitalgroup-form script
## What's being changed

We change the way we select the script tag that belongs to dotdigital forms that are embedded via our page builder. We now looking for script tags that contains `data-form-id` attribute.

## Why it's being changed

There is a client that uses this [module](https://amasty.com/facebook-pixel-for-magento-2.html) to inject facebook pixel in the magento store. Unfortunately this module looks like it interferes with our javascript implementation and prevents our code from selecting the correct script in the widget.js. As a result we can't get subscribers in subscribers table after they get subscribed via our dotdigital form.

## How to review / test this change

- Inject a new dd form via page builder to a frontend (attention you need to inject a new one, changes only apply to the new ones).
- Put a breakpoint to the widget.js subscribe to the form and confirm that the correct script selected.
- Make sure that the user subscribed to the newsletter_subscribers table.

Related work items: #196135, #196141