Merged PR 43633: Protect customer sync against failures in Enterprise customer plugin
## What's being changed

We're adjusting the logic and adding try/catch blocks in our CustomerPlugin to prevent errors surfaced here from blocking customer sync.

## Why it's being changed

Even though `Dotdigitalgroup_Enterprise` requires `Magento_Reward` and `Magento_CustomerSegment`, evidently some installations do not have certain tables belonging to these modules. In the event that e.g. the `magento_reward` table is missing, an exception is thrown in the plugin which prevents the syncing of any data for the supplied customer.

We've added the extra IF statements to allow merchants to switch off these mappings to resolve any errors that are surfaced.

There are separate try/catch blocks for reward-related data and segment-related data, to try and further confine the impact of specific missing tables.

## How to review / test this change

- Rename `magento_reward` to something different
- Map Enterprise data fields
- Run customer sync
- Ensure that other data fields are correctly exported for each customer
- Ensure that customer segment data is exported from the CustomerPlugin

Related work items: #204329