Merged PR 48356: Preserve newsletter subscribe function in forms if script is deferred
## What's being changed

Our Page Builder content type and the on-submit widget script.

## Why it's being changed

If scripts are deferred, we can't easily find the `data-add-respondent` flag that decides whether to natively subscribe the email address captured in the form. By placing this attribute on the `main` element we're not affected by script deferment.

Previously-embedded forms will still function, assuming javascript is not deferred (if it is they won't work anyway). The logic is like this:

- If the main container has the 'add-respondent' attribute (new form)
check if that attribute === 1

- If the main container does not have the 'add-respondent' attribute (old form)
check if the script's  'add-respondent' attribute === 1

## How to review / test this change

_Existing forms_
- On develop branch, embed a form (which must contain an email address) using our Page Builder content type with 'Add Respondent' set to 'Yes'
- View this on a page on the storefront
- Now switch branch to this one
- Clear cache
- Test your existing form
- Confirm that the email arrives in the newsletter_subscriber table

_New forms_
- Embed a new form using Page Builder, again with 'Add Respondent' set to 'Yes'
- On the front end, confirm that `data-add-respondent` is now added to the script tag's container div
- Submit the form
- Confirm that the email arrives in the newsletter_subscriber table

## Notes

It is possible to set deferred scripts using another plugin, but I don't think that's necessary to test this. If we're not looking for an attribute of the script, it doesn't matter where the script ends up in the DOM. I was able to spoof-defer the script by overriding the selectors e.g. `$(element).data('add-respondent-foo')` and `$(element).find('script[data-foo-id]')`.

Related work items: #227513